+++ PremiumScraps/Utils/ShipInventoryConditions.cs
@@ -18,7 +18,7 @@ namespace PremiumScraps.Utils
         {
             var item = player.currentlyHeldObjectServer;
             if ((item.itemProperties.name == "BombItem" && item is Bomb bomb && bomb.activated) ||
-                (item.itemProperties.name == "ControllerItem" && item is Controller controller && controller.isInControlMode) ||
+                (item.itemProperties.name == "ControllerItem" && item is Controller) ||
                 (item.itemProperties.name == "JobApplicationItem" && item is JobDark) ||
                 (item.itemProperties.name == "GazpachoItem" && item is SpanishDrink) ||
                 (item.itemProperties.name == "BookCustomItem" && item is StupidBook stupidbook && stupidbook.nbFinish != 0))